+++ tests/unit/test_finetuner.py
@@ -2,6 +2,7 @@ import docarray
 import pytest
 
 from finetuner.constants import CREATED, FAILED, FINISHED, STARTED, STATUS
+from finetuner.data import DATA_SYNTHESIS_EN
 
 
 @pytest.mark.parametrize(
@@ -35,13 +36,34 @@ def test_list_experiments(finetuner_mocker):
     'experiment_name',
     ['exp name', None],
 )
-def test_create_run(finetuner_mocker, experiment_name):
+def test_create_training_run(finetuner_mocker, experiment_name):
     data = docarray.DocumentArray().empty(1)
     run_name = 'run1'
     exp_name = experiment_name or 'default'
-    run = finetuner_mocker.create_run(
+    run = finetuner_mocker.create_training_run(
         model='resnet50',
         train_data=data,
+        run_name=run_name,
+        experiment_name=experiment_name,
+    )
+    assert run.name == run_name
+    assert run.status()[STATUS] in [CREATED, STARTED, FINISHED, FAILED]
+    assert run._experiment_name == exp_name
+
+
+@pytest.mark.parametrize(
+    'experiment_name',
+    ['exp name', None],
+)
+def test_create_synthesis_run(finetuner_mocker, experiment_name):
+    data = docarray.DocumentArray().empty(1)
+    run_name = 'run1'
+    exp_name = experiment_name or 'default'
+    run = finetuner_mocker.create_synthesis_run(
+        query_data=data,
+        corpus_data=data,
+        models=DATA_SYNTHESIS_EN,
+        num_relations=3,
         run_name=run_name,
         experiment_name=experiment_name,
     )